chapter_2001/src/test/java/ru/job4j/iterator/EvenLtTest.java
package ru.job4j.iterator;

import org.junit.Test;

import java.util.Iterator;

import static org.hamcrest.MatcherAssert.assertThat;
import static org.hamcrest.Matchers.is;

/**
 * Test.
 *
 * @author Sergey Gavrilov (mailto:sarmexin@gmail.com)
 * @version $Id$
 * @since 0.1
 */
public class EvenLtTest {
    private Iterator<Integer> it = new EvenLt(new int[]{4, 2, 1, 1});

    @Test
    public void testForEvenNumber() {
        assertThat(it.hasNext(), is(true));
        assertThat(it.next(), is(4));
        assertThat(it.hasNext(), is(true));
        assertThat(it.next(), is(2));
        assertThat(it.hasNext(), is(false));
    }

    @Test
    public void test2ForEvenNumber() {
        Iterator<Integer> it = new EvenLt(new int[]{4, 2, 1, 1, 2});
        assertThat(it.hasNext(), is(true));
        assertThat(it.next(), is(4));
        assertThat(it.hasNext(), is(true));
        assertThat(it.next(), is(2));
        assertThat(it.hasNext(), is(true));
        assertThat(it.next(), is(2));
        assertThat(it.hasNext(), is(false));
    }
}
